# src/cmd/http/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API Server",
        "title": "API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/publications": {
            "post": {
                "tags": [
                    "publications"
                ],
                "summary": "Publicar una publicacion",
                "parameters": [
                    {
                        "description": "objeto XD",
                        "name": "PublicationDto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.PublicationDtoResponse"
                        }
                    },
                    {
                        "description": "Imagenes[]",
                        "name": "images",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.GetPublicationResponse"
                        }
                    },
                    "503": {
                        "description": "Error con la base de datos",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/api/publications/username/{username}": {
            "get": {
                "tags": [
                    "publications"
                ],
                "summary": "Recibir publicaciones de un usuario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "N° page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.GetPublicationsResponse"
                        }
                    },
                    "403": {
                        "description": "Credenciales inválidas",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    },
                    "409": {
                        "description": "La sesión no existe. Probablemente porque la eliminaron",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    },
                    "503": {
                        "description": "Error con la base de datos",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/api/publications/{idPost}/like": {
            "get": {
                "tags": [
                    "publications"
                ],
                "summary": "Recibir el like de un usuario sobre una publicacion",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id de post",
                        "name": "idPost",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.GetLikeResponse"
                        }
                    },
                    "503": {
                        "description": "Error con la base de datos",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "publications"
                ],
                "summary": "Dar Like o dislike",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id de post",
                        "name": "idPost",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.GetLikeResponse"
                        }
                    },
                    "503": {
                        "description": "Error con la base de datos",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/api/publications/{idPublication}": {
            "delete": {
                "tags": [
                    "publications"
                ],
                "summary": "Eliminar una publicacion",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id de publicacion",
                        "name": "idPublication",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "503": {
                        "description": "Error con la base de datos",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/api/publications/{idPublication}/view": {
            "post": {
                "tags": [
                    "publications"
                ],
                "summary": "Añadir una \"view\" a una publicacion",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id de publicacion",
                        "name": "idPublication",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "503": {
                        "description": "Error con la base de datos",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.GetLikeResponse": {
            "type": "object",
            "properties": {
                "isLiked": {
                    "type": "boolean"
                }
            }
        },
        "controller.GetPublicationResponse": {
            "type": "object",
            "properties": {
                "publication": {
                    "$ref": "#/definitions/model.Publication"
                }
            }
        },
        "controller.GetPublicationsResponse": {
            "type": "object",
            "properties": {
                "publications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Publication"
                    }
                }
            }
        },
...
        "controller.PublicationDtoResponse": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 500
                },
                "idCategories": {
                    "description": "Images       []PublicationImageDto `form:\"images\" binding:\"dive\"`",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "model.Coord": {
            "type": "object",
            "properties": {
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                }
            }
        },
        "model.Image": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "mime": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.Profile": {
            "type": "object",
            "properties": {
                "avatar": {
                    "$ref": "#/definitions/model.Image"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "idUser": {
                    "type": "integer"
                },
                "likes": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                }
            }
        },
        "model.Publication": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "idProfile": {
                    "type": "integer"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Image"
                    }
                },
                "likes": {
                    "type": "integer"
                },
                "mentions": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "profile": {
                    "$ref": "#/definitions/model.Profile"
                },
                "tattoos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Tattoo"
                    }
                },
                "views": {
                    "type": "integer"
                }
            }
        },
        "model.Role": {
            "type": "string",
            "enum": [
                "user",
                "tattooArtist",
                "admin"
            ],
            "x-enum-varnames": [
                "USER_ROLE",
                "TATTOO_ARTIST_ROLE",
                "ADMIN_ROLE"
            ]
        },
        "model.Tattoo": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "coord": {
                    "$ref": "#/definitions/model.Coord"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "idPublication": {
                    "type": "integer"
                },
                "image": {
                    "$ref": "#/definitions/model.Image"
                },
                "likes": {
                    "type": "integer"
                },
                "profile": {
                    "$ref": "#/definitions/model.Profile"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Role"
                    }
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "utils.ErrorProblemDetails": {
            "type": "object",
            "properties": {
                "param": {
                    "type": "string",
                    "example": "3"
                },
                "pointer": {
                    "type": "string",
                    "example": "max"
                },
                "title": {
                    "type": "string",
                    "example": "name"
                }
            }
        },
        "utils.ProblemDetails": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Detalle técnico del error"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/utils.ErrorProblemDetails"
                    }
                },
                "title": {
                    "type": "string",
                    "example": "Descripción del problema para mostrar al usuario"
                },
                "type": {
                    "type": "string",
                    "example": "/docs/errors/errorPointer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "BearerJWTToken in Authorization Header",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}